backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "benchmark-eval-platform"
version = "0.1.0"
description = "FastAPI backend for math question model evaluation."
readme = "PROJECT_STATUS.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "psycopg[binary]>=3.2,<4.0",
    "alembic>=1.16,<2.0",
    "pydantic-settings>=2.8,<3.0",
    "httpx>=0.28,<1.0",
    "redis>=5.2,<6.0",
    "celery>=5.4,<6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
